refactor(app): migrate to nextjs
