+++ src/components/book-cover/book-cover.module.css
@@ -1,6 +1,3 @@
-@value colors: '../../styles/variables.css';
-@value primary-text-color, blur-background, grey-lite from colors;
-
 .cover {
   width: 30rem;
   height: 40rem;
