--- src/components/book-info/book-info.css
@@ -1,7 +1,0 @@
-
-.image {
-  margin-top: -6rem;
-  width: 300px;
-  height: 350px;
-  transition: all .35s;
-}
