+++ src/components/clipping-item/clipping-item.module.css
