+++ src/pages/_app.tsx
@@ -9,6 +9,7 @@ import { client, request } from '../services/ajax'
 import store from '../store/index'
 
 import 'react-toastify/dist/ReactToastify.css'
+import 'react-phone-input-2/lib/style.css'
 import 'emoji-mart/css/emoji-mart.css'
 import '../styles/global.css'
 import '../styles/tailwind.css'
