+++ src/pages/clipping/clipping.module.css
@@ -1,5 +1,3 @@
-@value colors: '../../styles/variables.css';
-@value primary-text-color, blur-background from colors;
 .clipping {
   min-height: 100%;
 }
@@ -29,7 +27,7 @@
   box-sizing: border-box;
   width: 100%;
   display: flex;
-  background-color: blur-background;
+  background-color: rgba(33, 150, 243, 0.9);
   margin: 0;
   cursor: pointer;
 }
