+++ src/pages/settings/settings.tsx
@@ -1,8 +1,6 @@
-import React, { useCallback, useEffect, useState } from 'react'
+import React from 'react'
 import Select from 'react-select'
-// import { DarkModeSwitch } from 'react-toggle-dark-mode'
 import { useTranslation } from 'react-i18next'
-import DarkSwitcher from './dark-switcher'
 import SimpleSwitcher from './simple-switcher'
 import Exports from './exports'
 import WebHooks from './webhooks'
