+++ src/pages/settings/webhooks.tsx
@@ -12,7 +12,7 @@ import FieldInput from '../../components/input'
 import { useFormik } from 'formik'
 import { toast } from 'react-toastify'
 import * as Yup from 'yup'
-import { ExportDestination, WebHookStep } from '../../../__generated__/globalTypes'
+import { WebHookStep } from '../../../__generated__/globalTypes'
 import { createNewWebHook, createNewWebHookVariables } from '../../schema/mutations/__generated__/createNewWebHook'
 import { deleteAWebHook, deleteAWebHookVariables } from '../../schema/mutations/__generated__/deleteAWebHook'
 import { useTranslation } from 'react-i18next'
