+++ src/pages/uploader/loading-modal.tsx
@@ -1,8 +1,7 @@
 import { AnimateOnChange } from '@nearform/react-animation'
-import React, { useEffect, useState } from 'react'
+import React from 'react'
 import { UploadStep } from './types'
 
-import './modal.styl'
 import { useTranslation } from 'react-i18next'
 
 type LoadingModalProps = {
@@ -25,7 +24,7 @@ function LoadingModal(props: LoadingModalProps) {
 
   return (
     <div
-      className='fixed top-0 left-0 right-0 bottom-0 flex justify-center items-center anna-mask z-50 anna-fade-in'
+      className='fixed top-0 left-0 right-0 bottom-0 flex justify-center items-center z-50 bg-black bg-opacity-75 backdrop-blur-lg anna-fade-in'
     >
       <div
         className='px-32 py-16 rounded shadow-lg bg-gray-100 flex justify-center items-center flex-col'
