--- src/pages/uploader/modal.styl
@@ -1,3 +1,0 @@
-.anna-mask
-  background-color rgba(0,0,0, .75)
-  backdrop-filter blur(5px)
